requirements.txt
flask~=3.0.3
kafka-python